hearts/tests: split off test_hand
