+++ hearts/tests/test_hand.py
@@ -1,0 +1,34 @@
+import pytest
+
+from hearts.hearts import Card
+from hearts.hearts import Hand
+
+
+def test_hand():
+    hand = Hand([
+        Card('A', 'h'),
+        Card('7', 'd'),
+        Card('6', 'h'),
+        Card('2', 's')
+    ])
+    assert Card('7', 'd') in hand
+    assert hand.has_suit('h')
+    assert not hand.has_suit('c')
+    with pytest.raises(ValueError):
+        hand = Hand([Card('d', '6')])
+    assert not hand.is_only_hearts()
+    hand = Hand([Card('A', 'h'), Card('7', 'h'), Card('2', 'h')])
+    assert hand.is_only_hearts()
+
+
+def test_hand_sort():
+    test = Hand.deserialize(['2d', 'Ts', '2c', 'Kh', 'Jc', '6s', 'Ac'])
+    test.hand_sort()
+
+    expected = Hand.deserialize(['2c', 'Jc', 'Ac', '2d', 'Kh', '6s', 'Ts'])
+    assert test == expected
+
+
+def test_serialize_hand():
+    hand = Hand([Card('2', 'c'), Card('3', 'h'), Card('4', 's'), Card('Q', 'h')])
+    assert Hand.deserialize(hand.serialize()) == hand
