+++ hearts/tests/test_trick.py
@@ -1,7 +1,6 @@
 import pytest
 
 from hearts.hearts import Card
-from hearts.hearts import Hand
 from hearts.hearts import Player
 from hearts.hearts import Trick
 
@@ -9,7 +8,6 @@ P1 = Player('Lauren')
 P2 = Player('Erin')
 P3 = Player('Jeremy')
 P4 = Player('Daniel')
-PLAYER_LIST = [P1, P2, P3, P4]
 
 """ Card class tests """
 
@@ -27,39 +25,6 @@ def test_card_validate():
         Card.deserialize('Q4')
     with pytest.raises(ValueError):
         Card.deserialize('2t')
-
-
-""" Hand Class Tests """
-
-
-def test_hand():
-    hand = Hand([
-        Card('A', 'h'),
-        Card('7', 'd'),
-        Card('6', 'h'),
-        Card('2', 's')
-    ])
-    assert Card('7', 'd') in hand
-    assert hand.has_suit('h')
-    assert not hand.has_suit('c')
-    with pytest.raises(ValueError):
-        hand = Hand([Card('d', '6')])
-    assert not hand.is_only_hearts()
-    hand = Hand([Card('A', 'h'), Card('7', 'h'), Card('2', 'h')])
-    assert hand.is_only_hearts()
-
-
-def test_hand_sort():
-    test = Hand.deserialize(['2d', 'Ts', '2c', 'Kh', 'Jc', '6s', 'Ac'])
-    test.hand_sort()
-
-    expected = Hand.deserialize(['2c', 'Jc', 'Ac', '2d', 'Kh', '6s', 'Ts'])
-    assert test == expected
-
-
-def test_serialize_hand():
-    hand = Hand([Card('2', 'c'), Card('3', 'h'), Card('4', 's'), Card('Q', 'h')])
-    assert Hand.deserialize(hand.serialize()) == hand
 
 
 """Trick Class Tests"""
